+++ CHANGELOG.md
@@ -82,6 +82,7 @@ with the `YY.MINOR.MICRO` scheme.
 - Fixed metadata loss in queries with LIMIT clause in picodata admin.
 - Fixed invalid volatile flag for rust-implemented builtin functions.
 - Fixed governor's `ConfigureReplication` step was broken during upgrade from before 25.5.3
+- Fixed that instances from tiers with can_vote=false attempting to promote to raft leader.
 
 
 ## [25.5.5] - 2026-01-26